Implemented the first version of the current activities. Currently there are two list of activities stack one after another. One of the hold the activities that the user owns and the other one holds the activities that the user is participating in.
When the method for loading activities changes there is going to
have to be some work done on the back end code since it heavily
relies on the code for the standard feed.

Furthermore the layout needs a lot of work since it looks a bit
shit but tbh I don't feel that it is not to important right now.

Signed-off-by: MiztaOak <ek.johan98@gmail.com>